Refactor: moving API models under new model dir
