+++ tests/test_schedule.py
@@ -4,7 +4,7 @@ import pytest
 import yaml
 from kubernetes.client import ApiException
 
-from openshift_resources.velero.schedule import Schedule
+from openshift_resources.model.velero.schedule import Schedule
 
 path = str(Path(__file__).parent)
